+++ src/main/java/net/jrodolfo/java_evolution/java04/security/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for Java 4 integrated security APIs.
+ * Executable local examples for Java 4 integrated security APIs.
  */
 package net.jrodolfo.java_evolution.java04.security;